+++ src/Traits/Metable.php
@@ -95,4 +95,13 @@ trait Metable
 
         return json_encode($value);
     }
+
+    /**
+     * Get an attribute from the model.
+     *
+     * @param  string  $key
+     *
+     * @return mixed
+     */
+    abstract public function getAttribute($key);
 }