+++ src/User.php
@@ -117,7 +117,7 @@ class User extends Eloquent implements Authorizable, UserContract
     /**
      * Activate current user.
      *
-     * @return \Orchestra\Model\User
+     * @return $this
      */
     public function activate()
     {
@@ -129,19 +129,23 @@ class User extends Eloquent implements Authorizable, UserContract
     /**
      * Assign role to user.
      *
-     * @param  int|array  $roles
+     * @param  \Orchestra\Model\Role|int|array  $roles
      *
      * @return void
      */
     public function attachRole($roles)
     {
+        if ($roles instanceof Role) {
+            $roles = [$roles->getKey()];
+        }
+
         $this->roles()->sync((array) $roles, false);
     }
 
     /**
      * Deactivate current user.
      *
-     * @return \Orchestra\Model\User
+     * @return $this
      */
     public function deactivate()
     {
@@ -153,12 +157,16 @@ class User extends Eloquent implements Authorizable, UserContract
     /**
      * Un-assign role from user.
      *
-     * @param  int|array  $roles
+     * @param  \Orchestra\Model\Role|int|array  $roles
      *
      * @return void
      */
     public function detachRole($roles)
     {
+        if ($roles instanceof Role) {
+            $roles = [$roles->getKey()];
+        }
+
         $this->roles()->detach((array) $roles);
     }
 
@@ -273,7 +281,7 @@ class User extends Eloquent implements Authorizable, UserContract
     /**
      * Suspend current user.
      *
-     * @return \Orchestra\Model\User
+     * @return $this
      */
     public function suspend()
     {