+++ docs/changes.md
@@ -7,7 +7,7 @@ title: Model Change Log
 
 ### v3.3.1 {#v3-3-1}
 
-* Allow `Orchestra\Model\Traits\OwnedByTrait` and `Orchestra\Model\Traits\OwnsTrait` to accept `null` instead of just `Illuminate\Database\Eloquent\Model`.
+* Allow `Orchestra\Model\Traits\OwnedBy` and `Orchestra\Model\Traits\Owns` to accept `null` instead of just `Illuminate\Database\Eloquent\Model`.
 
 ### v3.3.0 {#v3-3-0}